backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
bech32==1.2.0
bitarray==3.8.0
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.0
cosmpy==0.11.2
cytoolz==1.1.0
distlib==0.4.0
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.120.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
frozenlist==1.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hexbytes==1.3.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
parsimonious==0.10.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==5.29.5
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
PyNaCl==1.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyunormalize==17.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.28.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.48.0
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
uagents==0.22.10
uagents-core==0.3.11
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.35.4
web3==7.13.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.22.0
